+++ helpers/helpers_xxx_console.js
@@ -11,10 +11,10 @@ let conLogMaxSize = 5000000; // File size, in bytes. Setting to zero or null dis
 function consoleLog() {
 	let log = '';
 	// Load previous log
-	if (utils.IsFile(conLog)) {log += utils.ReadTextFile(conLog, convertCharsetToCodepage('UTF-8'));}
+	if (utils.IsFile(conLog)) {try {log += utils.ReadTextFile(conLog, convertCharsetToCodepage('UTF-8'));} catch (e) {}}
 	// Add HH:MM:SS
 	const stamp = '[' + new Date().toLocaleTimeString() + ']';
-	log += (log.length ? '\n' : '') + stamp;
+	log += (log && log.length ? '\n' : '') + stamp;
 	// Unpack args
 	[...arguments].forEach((arg, i) => {
 		const type = typeof arg;